backend/requirements.txt
flask
flask-cors
openai-whisper
librosa
pydub
scikit-learn
numpy
scipy
